refactor: update MyService package structure
