+++ app-service-hilt-inject/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
             </intent-filter>
         </activity>
 
-        <service android:name=".MyService" />
+        <service android:name=".service.MyService" />
     </application>
 
 </manifest>
