+++ app-service-hilt-inject/src/main/java/com/github/only52607/compose/window/hilt/inject/MainActivity.kt
@@ -25,6 +25,7 @@ import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.github.only52607.compose.window.checkOverlayPermission
 import com.github.only52607.compose.window.hilt.inject.repository.UserPreferencesRepository
+import com.github.only52607.compose.window.hilt.inject.service.MyService
 import com.github.only52607.compose.window.hilt.inject.ui.DialogPermission
 import com.github.only52607.compose.window.hilt.inject.ui.theme.ComposeFloatingWindowTheme
 import dagger.hilt.android.AndroidEntryPoint
