+++ app-service-hilt-inject/src/main/java/com/github/only52607/compose/window/hilt/inject/service/MyService.kt
@@ -1,4 +1,4 @@
-package com.github.only52607.compose.window.hilt.inject
+package com.github.only52607.compose.window.hilt.inject.service
 
 import android.app.Service
 import android.content.Context
